# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "admin登陆",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Info",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.AdminUserLoginInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.AdminUserLoginInfo": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "api.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ResultCode"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "api.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                -1
            ],
            "x-enum-varnames": [
                "Success",
                "Error"
            ]
        },
...
    }
}